(MAINT) Don't print container messages to CLI
Silence the container entrypoint echo statements. Since this is a CLI
container and not a service container, it should not have these verbose
logging style statements being printed out. They just get in the way of
the user trying to run the CLI and see the output of their command.

Since this is breaking consistency with our other containers (which all
happen to be service containers and not CLI containers) I left comments
in place saying why we're not printing out any messages.